improve dropdowns
